+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.60.4
+- fix: check only for non-matching property when the property name is not quoted.
+  - quoted property name in `style!` macro will be an escape hatch for few cases such as the use of prefixes, ie: "-webkit, -ms, -moz"
+
 ## 0.60.3
 - return `Option` for style property lookup instead of panicking.
 - Make `style!` provides a helpful error when property names used is not found in the lookup.